+++ test/babel.config.js
@@ -1,0 +1,7 @@
+module.exports = {
+  presets: [[ '@babel/preset-env', {
+    bugfixes: true,
+    corejs: 3,
+    useBuiltIns: 'usage'
+  }]]
+};