+++ test/rollup.config.js
@@ -1,0 +1,11 @@
+export default {
+  input: '../',
+  plugins: [
+    'eslint',
+    'buble'
+  ],
+  output: {
+    file: 'bundle.js',
+    format: 'esm'
+  }
+};